Still need to do LOS, StabilityDefenses,     "Cost": 7100000,     "Rarity": 5, Chassis UIName Details Hardpoints BattleValue calcuations LOSSourcePositions LOSTargetPositions   "ChassisTags": {     "items": [],     "tagSetSourceFile": ""

+++ Code/Utils/C#/BattleEngineJsonCreation/BattleEngineJsonCreation/BattleTechJsonClass.cs
@@ -350,7 +350,7 @@ namespace BattleEngineJsonCreation
         public long? SimGameMechPartCost { get; set; }
 
         [JsonProperty("Version", NullValueHandling = NullValueHandling.Ignore)]
-        public long? Version { get; set; }
+        public double? Version { get; set; }
 
         [JsonProperty("Locations", NullValueHandling = NullValueHandling.Ignore)]
         public List<MechDefLocation> Locations { get; set; }
